requirements.txt
fastapi
uvicorn
python-telegram-bot==20.3
pymediainfo
Pillow
piexif